third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/import-attributes/invalid-type-attribute-error-expected.txt

This is a testharness.js-based test.
FAIL Test that invalid module type attribute leads to TypeError on window. assert_equals: expected function "function TypeError() { [native code] }" but got function "function SyntaxError() { [native code] }"
Harness: the test ran to completion.
